refactor: sonarqube fix for maintainability
