+++ back/healthapp/src/main/java/com/healthcare/infrastructure/security/controller/AuthController.java
@@ -1,8 +1,10 @@
-package com.healthcare.infrastructure.security.service;
+package com.healthcare.infrastructure.security.controller;
 
+import com.healthcare.infrastructure.dto.response.ResponseTokenDTO;
+import com.healthcare.infrastructure.security.service.AuthService;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -14,18 +16,17 @@ import com.healthcare.infrastructure.meta.DocumentedResponse;
 
 import jakarta.validation.Valid;
 
+@RequiredArgsConstructor
 @RestController
 @RequestMapping("/auth")
 public class AuthController {
 
-    @Autowired
-    private AuthService authService;
+    private final AuthService authService;
 
-    @DocumentedResponse(responseCode = "200",
-    description = "Respuesta de login correcta",
+    @DocumentedResponse(description = "Respuesta de login correcta",
     content = @Content(schema = @Schema(example = "{\"token\": \"jwt_token\"}")))
     @PostMapping("/login")
-    public ResponseEntity<?> loggedIn(@RequestBody @Valid RequestLoginDTO login) {
+    public ResponseEntity<ResponseTokenDTO> loggedIn(@RequestBody @Valid RequestLoginDTO login) {
         return authService.loginUser(login);
     }
